Memfault Firmware SDK 0.4.2 (Build 90228)

+++ components/core/src/memfault_sdk_assert.c
@@ -19,7 +19,7 @@ void memfault_sdk_assert_func(void) {
   void *return_address;
   MEMFAULT_GET_LR(return_address);
 
-  MEMFAULT_LOG_ERROR("ASSERT! LR: 0x%x", (int)return_address);
+  MEMFAULT_LOG_ERROR("ASSERT! LR: 0x%x", (int)(uintptr_t)return_address);
   memfault_platform_halt_if_debugging();
   memfault_sdk_assert_func_noreturn();
 }
